Improved the Let syntax.
At the expense of a wordier declaration, the general usage is simplified.

The 'preserve' invocation has also been removed. The environment is always
preserved when defining new variables and can only be explicitly cleared by
calling Let.clear() or be instantiating a new environment when a new one is
needed.