Release version 1.0.0-beta
Given was originally called LazyLet, and as LazyLet it has been used for testing
production code at my day job for a number of weeks with no major issues.

There was not only a rename - there has been a fairly significant refactoring
so I'll release a beta now and after a couple of weeks of further production
use without new bugs I'll release a 1.0.0.