Removed the immediate form
The immediate form was the ability to define a value non-lazily, like this:

```coffeescript
    given foo: 'bar'
```

instead of:

```coffeescript
    given foo: -> 'bar'
```

Accidentally omitting the function would cause *immediate* evaluation of the
variable which has proven to be a trap for the unwary at my day job, and being
able to support both forms offers zero advantage.

Given will now emit a meaningful error when forgetting to wrap the RHS in
a function.